Fix TypeError race condition in htex draining code
Before this PR, sometimes draining would fail with this error message:

Traceback (most recent call last):
  File "/home/runner/work/parsl/parsl/.venv/lib/python3.10/site-packages/parsl/p
rocess_loggers.py", line 27, in wrapped
    r = func(*args, **kwargs)
  File "/home/runner/work/parsl/parsl/.venv/bin/process_worker_pool.py", line 32
2, in pull_tasks
    next_interesting_event_time = min(last_beat + self.heartbeat_period,
TypeError: '<' not supported between instances of 'NoneType' and 'float'

because the code incorrect sets drain time to None to indicate that further
draining messages should not be sent. An earlier version of the drain code used
None in that situation, but this was changed to +inf in the version that was
actually merged.

This error is statically detectable by mypy, if you turn on mypy checking for
pull_tasks, although it also raises other type errors.