Support context manager protocol in DFK (#3260)
Recently, PR #3165 removed at-exit cleanup of DFKs, so after that PR, a user has to explicitly invoke the DFK cleanup method. More justification for this is in issue #3104

This PR adds context manager support to the DFK, so that it can be used in a with block around an entire workflow.