Inline fail_job_async
This should not change any behaviour.

The call happens only in one place, and it only modifies a block-related data
structure, right next to code that more directly modifies block-related data
strucutres.

This PR makes all of those block-related data structure changes happen at the
same level.

This call is not asynchronous, and this PR also removes that confusing use of
the term async - there is a historical reason,