Return DFK in context manager (#3296)
When attempting to use the with statement with parsl.load() to manage a DataFlowKernel object (dfk). Currently, the context manager protocol does not correctly assign the loaded DataFlowKernel to dfk, resulting in dfk being None. The desired solution is to ensure that with parsl.load() as dfk: That assigns the loaded DataFlowKernel object to dfk