Scale in blocks at shutdown using Job Status Poller (#3302)
This will now scale in blocks using the job status poller scale in code,
which means the (moved) shutdown scale in loop does not need to send its own
BLOCK_INFO messages.

Behaviour-wise:

The scale_in call is changed from the executor version to the status poller
version. That latter version is mostly a passthrough to the executor version,
with the above-mentioned monitoring code change. So the behaviour of the
scale in call + monitoring should be the same.

The parameter to the scale_in call is a count of blocks. This is intended to
mean "all of them" - by counting however many blocks are known and using that
count.

Prior to this PR, the number of known blocks was the number of jobs in the
provider's `resources` dictionary - a structure that should otherwise be
private to the particular provider.

Post this PR, the number of known blocks is the number of blocks known to the
JobStatusPoller module: this is a possibly different number, although it
should be at least the number of blocks which have been submitted (either by
getting a new status in PollItem.scale_out or by being present in the most
recent provider cache refresh in PollItem.poll).

Because it should be at least the number of submitted blocks, this number
should be large enough to represent "all of them" (with the above meaning).