Add ZipFileStaging to archive outputs into a zip file (#3319)
This output file staging mechanism provides a zip: URL scheme, naming
a file inside a zip file; output files from apps can be staged out into
this location. Input staging is not supported.

The use-case for this is as part of a sequence of PRs separating
PR #3306, leading to stdout staging into zipfiles.