Remove issue363 test marker
This marker originally referred to issue #363 for tests which should not be
expected to pass if running in an environment where stderr and stdout are not
staged back to the submitting system by the executor.

Since it was created, it then expanded in use to more generally refer to any
test which had a problem in environments which did not properly make either
output files or stdout/stderr available on the submitting system, as a
separate behaviour from the core of parsl behaving properly. An example of
that would be running in a multi-site environment without working file
staging.

However, since all that happened, the feel in the Parsl community is that
running with a shared filesystem is the core supported usage mode of Parsl;
and also since then, notions of file staging have become much stronger,
both in Parsl core and in individual executors, making it a much more
reasonable test requirement that if you make a file, it can be expected to
be available for the test to inspect - and that if you are in an environment
that does not support that, you should not expect the test suite (or Parsl)
to work.

Because of all of that, this PR removes the issue363 test marker entirely.
It was removed from Work Queue and Task Vine test environments in PR #3327,
leaving only the Radical-PILOT test environment setting that.

It turns out only two tests don't work in the Radical-PILOT environment
when issue363 is removed:

One requires support for the slightly awkward "stdout/err opening modes,
specified by tuples" feature, and this PR labels and disables that test
using a new 'executor_supports_std_stream_tuples' label. (In general, an
executor which does staging of stdout/err streams is likely to not support
the features offered by std stream tuples, such as appending to stderr
files from multiple task - so this label is not Radical-PILOT specific)

Another test looks like it is broken exception handling in the Radical-PILOT
executor, and so it is labelled with a new issue specific label, relating to
the bug report for that issue - issue #3328, issue3328.