Move monitoring radio reference from JobStatusPoller to DFK
This is a step in consolidating responsibility for BLOCK monitoring into
the BlockProviderExecutor.

Before this PR, executors were already configured ad-hoc with a lot of
monitoring information. This PR adds another attribute to that collection
for the local monitoring radio.

This PR removes some more state from the PolledExecutorFacade class.

This PR should not change behaviour - it changes the code path by which the
job poller's monitoring radio is found.