Fix scale_in return values for Work Queue and Task Vine
Before this PR, the scale in methods for those executors returned None,
in violation of the type specification in BlockProviderExecutor.
This PR adds type annotations to cause those methods to be type checked.

Tihs PR adds handling for return value based on the HTEX implementation
of scale_in to both Work Queue and Task Vine executors.

This PR adds a few debug logs around behaviour there.

These return values are not used anywhere (which is why this has not been
a problem) but an upcoming PR will both make use of them (in monitoring)
and test that something is returned.