Fix typing on general strategy
This PR adds a type annotation to the function signature.

That has the knock-on effect of enabling type checking within the
method, revealing two mypy errors.

One is unreachable code: because the strategy will only be called with a
BlockProviderExecutor executor, it does not need to check if it has been
called with a BlockProviderExecutor.

The other type error comes from mypy not realising that a path is
unreachable: the idle time of an executor may be None, but at the point
of the type error, an `if` statement has ensured that it is never none.
This PR adds an assert statement to assert that assertion to the type
checker (and, incidentally, to the logs via wrap_with_logs).