Unabstract executor.create_monitoring_info (#3339)
This method makes sense for BlockProviderExecutors, where it is used
by the JobStatusPoller. It does not make sense for ParslExecutors in
general and is not otherwise used. So remove from the ParslExecutor
base class, which removes the dependency of parsl.executors.base
on parsl.jobs.*, a cleaner separation of concerns.

This method is implemented in HighThroughputExecutor, but the information
generated only uses BlockProviderExecutor information and would make
sense for other BlockProviderExecutors (Work Queue, Task Vine and the
erstwhile IPP and Extreme Scale executors). Using this implementation
has been prototyped in the desc branch of parsl since 2021. So move that
implementation out of out of HighThroughputExecutor and into
 BlockProviderExecutor.

This should not change any behaviour when using HighThroughputExecutor.
When using Work Queue or Task Vine, this change will bring block
monitoring support to those executors. As noted above, this has been
used successfully with Work Queue since 2021 in the desc branch.
This PR modifies the basic monitoring test to test against those executors
in addition to the HighThroughputExecutor.

This PR modifies the type annotations of create_monitoring_info to type
check properly - the previous htex implementation was not type-checked.
The return type is modified to a sequence, because List in invariant on
its parameter type, which is incompatible with adding Dict[str, Any]
to it.