Rework monitoring recording of invalid stdout/err streams (#3344)
This PR stops treating None as an invalid stdout/err: it is a legitimate
value for stdout/err. Prior to this PR, using None was resulting in
the stdout/err being recorded in monitoring as a stringified exception,
like this:

```
sqlite> select task_stderr from task;
type of argument "stdfspec" must be one of (os.PathLike, str, Tuple[str, str], Tuple[os.PathLike, str]); got NoneType instead
```

Before this PR, an invalid stdout/err stream was logged at WARNING level
and the filename as recorded in the monitoring database was set to the
exception message.

This PR elevates that log to an ERROR, because this is an error,
including the exception message to give a debugging hint.

In this error case, a blank stdout or stderr will be recorded in the
monitoring database. Generally the task will fail later due to other
parts of the code being unable to interpret the specification.

This PR adds a few checks around stdout/err tests that should not have
any error raised - this test checks for ERROR rather than WARNING because
Parsl will legitimately output WARNING messages in some configurations
(such as when checkpointing is enabled but Parsl notices it hasn't
checkpointed anything)

Co-authored-by: Kevin Hunter Kesling <kevin@globus.org>