Add more cases for std stream logging (#3347)
Previously, stdout/err logs not output if the streams were not set
(i.e. defaulting to None) and logged with the entire stream target
if set.

This PR adds more cases here with case specific logging: if stdout/err
is not redirected, that is now logged. If tuple mode is used, the
tuple is unpacked and logged as separate fields. If an unknown format
is specified an error is logged (but no exception is raised - it is
not the job of this code to validate the stdout/err specifications)

This PR is in preparation for a new case to be added in an upcoming
PR which requires more serious formatting that was not accomodated
by the previous implementation.