Rename PolledExecutorFacade methods that overlap BlockProviderExecutor (#3351)
This is in preparation for moving all the PolledExecutorFacade code into
BlockProviderExecutor, in an upcoming PR.

There should now be no intersection in the attributes or methods of
PolledExecutorFacade and BlockProviderExecutor, in the methods that
will move. __init__ is an exception because it will merge with the
__init__ of BlockProviderException.

The renamed methods are slightly awkwardly named because there isn't a
short adjective that I have found that adequately describes the difference
between BlockProviderExecutor and PolledExecutorFacade behaviours.

Future untangling work in this area will necessarily involve quite a lot
of rearranging and picking appropriate names for whatever remains is part
of that future work - so no bikeshedding over adjectives.

This PR should only rename methods. It should not change any behaviour.