Fix two CI-rot breakages (#3360)
Two simultaneous CI-rot breakages have happened and this PR fixes them simultaneously - they cannot be fixed individually because neither fix is sufficient to pass CI.

RADICAL-Pilot installation broke recently in CI, through some change outside of the Parsl repository, and even with a pinned RADICAL-Pilot version. More recent radical-pilot will install OK, and this PR advances the pin to that more recent version. This might be something to do with this fix introduced in radical-cybertools/radical.pilot#3169 but I'm not completely clear.

Issue #3037 started happening in CI. That issue is that a test checks a property that it might not actually be reasonable to expect to hold. This PR skips that issue, leaving further work for whoever adopts issue #3037.