Treat stdout and stderr as stageable Files (#3348)
Prior to this PR, stdout and stderr were treated as string pathnames interpreted in an executor specific context: for example, htex treats them as paths on the worker filesystem; Task Vine treats these paths as names of files that should be staged to the submitting system.

This is/was different to the treatment of files placed in the outputs kwarg of an app using File objects. Those file objects can be staged out arbitrarily using Parsl file staging.

This PR allows File objects to be specified for bash_app stdout and stderr kwargs, and stages those files out in the same way as outputs-kwarg files are staged.

This PR preserves the previous str-based behaviour without staging. This is inconsistent with the data staging mechanism, which only accepts Files, but preserves backwards compatibility, which is important here because stdout and stderr kwargs are used by users a lot.

When this new staging mechanism is used, DataFutures for stdout/err Files are made available to the user via app_future.stdout and app_future.stderr. This properties existed previously, and when using strings for stdout/err, retain their previous behaviour of returning those strings.

This mode is incompatible with tuple-based stdout/err specifications, which allow opening stdout/err with specific file modes, and rely on the user to manage those files outwith Parsl: that concept is not compatible with the file staging model of "write a file once, don't ever modify it".

Fixes issue #363