Fix broken test that cannot be run multiple times in same environment.
This test used to re-use a temporary intermediate file by appending to it.
When run multiple times, it failed with errors like this as the results
from previous runs are also observed by the test:

```
E               AssertionError: assert '2\n2\n2\n2\n' == '2\n'
```

This PR uses tmpd_cwd to avoid that behaviour.