Fix stdout/err log errors introduced by PR #3347 (#3379)
PR #3347 attempted to do case-based logging of all the different kinds
of stdout/err specification.

It failed to capture some of the cases involving os.PathLike, and so
after PR #3347, those cases would log an ERROR that the specification
was unknown. This new behavior is only a new ERROR logging message -
PR #3347 did not change other behaviour.

This PR also amends a rich test of stdout/err specification types
introduced in PR #3363 to check that no ERROR messages are logged during
these tests.