Fix typo in test invocation that results in skipped DFK cleanup
This typo is not a syntax error, although it might superficially look
like one: it merely computes the method that if invoked, would clean
up the DFK, and then discards the result of that computation rather
than invoking it.

This typo resulted in an entire DFK being left behind after the test
finishes.