Add logs around usage tracker shutdown
Parsl shutdown logs are deliberately rich at shutdown time because this is an
incredibly hang prone part of the code-base.
See https://github.com/Parsl/parsl/labels/safe-exit

This PR adds logs bracketing the shutdown of the usage tracker, in line
with later shutdowns of the job status poller, executors and monitoring.