Shut down usage tracking harder
This invokes .close() on usage tracking, which releases two fds per
sub-process (so 4 for each usage-tracked DFK = 2 x {start, stop})
more immediately.

This PR reworks logging around shutdown to emphasise that something
is wrong if SIGKILL is needed, and to add a debug line right before
a potentially long delay.