.close() interchange process to release more fds
This releases two fds in the workflow process used to communicate with
the interchange process, and is part of work to release more resources
explicitly at shutdown rather than leaving them until process end.

One test was making an out-of-API shutdown of HTEX on the basis that
the DFK would not shutdown this executor when a bad state was set. That
used to be true, but PR 2855, which re-arranged some shutdown behaviour,
(accidentally?) changed shutdown to always happen. That is, I think,
the right shutdown behaviour: if an executor wants to suppress parts of
its own shutdown based on internal state, that's the business of the
executor, not the DFK.

This PR removes that out-of-API shutdown.