close channels!
its architecturally a bit ambiguous who should be responsible for this... channels live inside providers inside executors and its arguable that providers shoudl do that, when closed by executors (which they are not).

however, the DFK contains a bunch of channel initialization, and so it is legitimate for it to also contain a bunch of channel cleanup

TESTING: nothing actually tests this - perhaps a mock channel that we check DFK setup and shutdown works on?

I have seent this test in CI on cleanup of:

parsl/tests/test_htex/test_connected_blocks.py

E               AssertionError: The provider model assumes a provider has channel(s)
1986

but that test works for me...
?