close in test... but doesn't close in DFK ... so that needs to happen too (and be tested) (see different patch)
it should be the responsibility of the DFK to close channels (at least in that other patch)
and so that's why this test, which avoids the DFK, needs to close channels itself

but looks like channel.close() doesn't close down threads enough?
I can see it now closes the transport - the transport thread is now reported as unconnected,
prior to calling close it was reporting itself as connected. but how to make the transport
thread go away entirely?

it looks like the thread goes away "eventually" - a 20s sleep for example, is long enough
- so how to do that with joins?