Close SSH channel after test
This is paired with DFK behaviour introduced in PR #3503 which closes channels
in the DFK. This PR modifies a test which does not use the DFK to also perform
that shutdown.

With test parsl/tests/test_providers/test_local_provider.py, this reduces the
number of threads remaining at the end of execution from 3 to 2.