Factor interchange monitoring code into a ZMQRadioSender
From an interchange perspective: this is a refactoring intended to clarify
that the interchange isn't doing anything special wrt. monitoring
messages and that it can send monitoring messages in the same way that
remote workers can.

From a monitoring perspective: this pulls ZMQ sender code out of the
interchange and puts it in a place that is more natural for ongoing
development. For example, a potential future use with Work Queue and
Task Vine is that workers would also benefit from using ZMQ to send
monitoring messages.

In some potential use cases, it might be desirable to configure the
radio used by the interchange instead of the hard-coded ZMQRadio. On-going
work in draft PR #3315 addresses configuration of different types of radio
and that work should be relevant here too.