make interchange use the radio interface for forwarding monitoring messages
This is intended to clarify that what an interchange is sending is nothing special, beyond a normal monitoring message.

This is intended to loosen up future opportunities in two ways:

this is hard-coded / not configurable in the interchange - it might be later desirable to make
the interchange configurable if wanting to do other stuff with this status information
- eg christines prototype use case (see PR #3359) of wanting to run specific commands when
a worker connects (where a radio message is treated as a more general event hook),
or sending monitoring info to other places not the monitoring db

it might be desirable to deliver monitoring information from a worker node using ZMQ,
for example when using work queue co-processes or the task vine equivalent - where
the cost of starting a ZMQ connection might be amortised across many tasks.