Unregister atexit hook at cleanup
This hook is only to deal with reminder the user about cleanup and is
otherwise unnecessary.

The exit warning about cleanup state is now switched based on the
presence of the atexit handler or not, rather than the cleanup called
variable.

Effects of this PR:

* removed log noise

* one less garbage collector reference to the DFK - although that won't
directly lead to the DFK being collected earlier, it might in combination
with other future work.