Add exit choices for the DFK context manager
Most important is a ``wait`` exit mode which will make the context manager
wait for tasks to complete at exit, unless an exception is raised. This
addresses a common usability problem where users forget to wait for their
tasks and are then confused by their workflow exiting immediately, but
does not defer exceptions - this block-on-exception behaviour was the main
usability problem when attempting this before in PR #610.