Do parsl.clear() at DFK cleanup when relevant
It was UI wart that parsl.clear() is necessary for a user to do, arising from
a long-ago distinction between creating a DFK (of which potentially a user
might make many) and asking some other piece of Parsl to manage a single
default DFK (parsl.load). It has been an ongoing usability problem for users
to understand why they need to do both parsl.dfk().cleanup() and parsl.clear().

This PR makes a lot of changes throughout the test suite to remove parsl.clear
statements that are unnecessary after this PR - tested in development by a
removable assert.