Pin radical.utils version, to work with pinned radical.pilot
On 10th May 2024, a new version of radical.utils was released, which does not
work with the pinned radical.pilot==1.52.1 used by Parsl.

I also tried upgrading radical.pilot to 1.60, released alongside the new
radical.utils, but that does not work for other reasons documented in the
below issue.

See Parsl issue #3429