Remove return bool from Channel.close() (#3403)
Theres was no defined meaning to the return bool either in the code
or in the human text:

- no code calls close to give an example of interpretation

- it's not success/fail because local channel returns False
to indicate that it didn't need to do anything, not that there
was a failure.

Other .close() style methods return None and raise an exception
if there is a problem. This PR pushes Channel.close() in this
direction.

A separate PR will actually invoke this .close() method.