Do parsl.clear() at DFK cleanup when relevant (#3444)
It was UI wart that parsl.clear() is necessary for a user to do, arising from a long-ago distinction between creating a DFK (of which potentially a user might make many) and asking some other piece of Parsl to manage a single default DFK (parsl.load). It has been an ongoing usability problem for users to understand why they need to do both parsl.dfk().cleanup() and parsl.clear().

parsl.clear() will no longer be necessary - although it will be harmless to continue to do so.

This PR makes a lot of changes throughout the test suite to remove parsl.clear statements that are unnecessary after this PR (tested in development by a removed assert)

When invoking an app after the default DFK has been shut down, the error message will now be that a config has not been loaded (must first load config) rather than Cannot submit to a DFK that has been cleaned up - because the cleaned-up default DFK will have been unset, rather than left as default.