Add pluggable DependencyResolvers and an optional deep-structure resolver (#3111)
This PR allows users to change what it means to resolve (wait for and turn into a value) a dependency. The existing default behaviour is preserved in SHALLOW_DEPENDENCY_RESOLVER. This PR adds a plugin DEEP_DEPENDENCY_RESOLVER which knows how to walk over trees of several Python compound types (such as lists and dicts) looking for Futures.