Use sacct to get slurm job information (#3422)
Changes from the slurm squeue command to the sacct command. The sacct command is a bit easier on the Slurm scheduler as it connects to the slurm database instead of the slurm controller. One other part of using sacct is you can get job information for jobs which has finished as well so there may not be a need for the jobs_missing checks that are currently in the code.

Co-authored-by: Ben Clifford <benc@hawaga.org.uk>