fiddle with markings to deal with no shared fs and no staging
many of these newly marked "no shared fs" tests do not need to be shared fs, but instead could be staging-based?

it would be good to test the tests that aren't marked as staging required with no staging providers configured, even in normal CI, to validate that test mark is applied properly.

should also rewrite the mark name as shared_fs_required to match the X_required format?