Add a marker for tests which test errors caused by unix filesystem permissions
I've encountered a bunch of situations in the last few months where people
were running the test suite in environments where filesystem permissions were
not enforced - various container/virtual machine situations running as root.

In such a situation, some tests are not expected to pass, as those tests test
that a (not-happening) OS-level filesystem permission error is correctly
passed back to the Parsl user.