Launch tasks in from a shallow Python stack to avoid recursion errors (#3478)
Prior to this PR, task dependencies were launched from fairly arbitrary places in the call stack: for example, wherever a particular Future callback happens, sometimes deep inside executor-specific threads and/or in launch code for earlier tasks. In situations described in #3472 this could result in a a Python call stack overflow, as long call chains accumulate.

This PR makes all task launches be queued to happen from a near-empty stack in a single thread (managed as a concurrent.futures.ThreadPoolExcutor) - with a "invoke this call soon" pattern.

The launch_if_ready method was already intended to be launched multiple times from multiple threads. this PR might make the invocation to launch_if_ready happen a bit later, but correctness-wise that should be fine: a task can only become more ready to run, not less ready.

This PR introduces a test test_dependency_deep, which simulates a situation where this #3472 failure happens (before this PR), and tests (for the future) that tasks are not launched beyond a fairly arbitrary stack depth.