Fixed Kubernetes worker container launch command to remove trailing semicolon (#3486)
The launch command is formatted with a newline at the end so when the trailing semicolon is added here, it causes the semicolon to be run as a separate bash command by the worker. A single semicolon is a syntax error in bash and produces an error in container logs when the container tries to run it. Removing it is safe since it's at the end of the command and nothing is concatenated after it.